backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.27.0
langchain==0.1.5
langchain-community==0.0.19
sentence-transformers==2.3.1
qdrant-client==1.7.0
beautifulsoup4==4.12.3
html-to-json==2.0.0
google-generativeai==0.3.2
python-multipart==0.0.9
jinja2==3.1.3
joblib==1.3.2
numpy==1.26.3
vinai-phobert==0.1.0
asyncio==3.4.3
aiofiles==23.2.1
fastapi-cache2==0.2.1
tenacity==8.2.3
loguru==0.7.2
pytest==7.4.4